desktop_app/requirements.txt
mysql-connector-python>=8.0
pillow>=10.0
bcrypt>=4.0
